python/requirements.txt
mysql-connector-python==8.0.26
faker==8.10.0
python-dotenv==0.19.0